modify alignhash flags

+++ test/test_bit.cpp
@@ -21,7 +21,9 @@ int main()
 	GRAYCODE_TO_BIN64(s);
 
 	if (s != r)
-		fprintf(stderr, "expected %016llx, acutal %016llx\n", r, s);
+		fprintf(stderr, "expected %016llx, acutal %016llx\n",
+			(unsigned long long)r,
+			(unsigned long long)s);
 	else
 		printf("passed\n");
